\name{lookup_codes}
\alias{lookup_codes}
\title{Look up information on specific JIS prefecture and
municipality codes.}
\usage{
lookup_codes(codes, date = Sys.Date())
}
\arguments{
  \item{codes}{A vector of JIS prefecture or municipality
  codes.}

  \item{date}{A Date object.}
}
\value{
A data frame like that returned by
\code{\link{lookup_names}}.
}
\description{
Look up information on specific JIS prefecture and
municipality codes.
}

