\name{lookup_details}
\alias{lookup_details}
\title{Look up detailed information on a singe JIS prefecture or
municipality code.}
\usage{
lookup_details(code)
}
\arguments{
  \item{code}{A single JIS prefecture or municipality
  code.}
}
\value{
A data frame with an extensive summary of the information
about the given locality code contained in the NSTAC
database.
}
\description{
Look up detailed information on a singe JIS prefecture or
municipality code.
}

