\name{prefecture_codes}
\alias{prefecture_codes}
\title{Get the JIS codes and names for all 47 Japanese prefectures
(retrieved from disk).}
\usage{
prefecture_codes()
}
\value{
A data frame containing codes and names for all
prefectures. Codes are trimmed to 2 digits.
}
\description{
Get the JIS codes and names for all 47 Japanese prefectures
(retrieved from disk).
}

